--- a/src/marvin/fns/run.py
+++ b/src/marvin/fns/run.py
@@ -15,19 +15,43 @@ from marvin.handlers.handlers import AsyncHandler, Handler
 T = TypeVar("T")
 
 
+def _tasks_are_independent(tasks: list[Task[Any]]) -> bool:
+    """Check if tasks have no dependencies between each other."""
+    for i, task1 in enumerate(tasks):
+        for j, task2 in enumerate(tasks):
+            if i != j:
+                # Check if task1 depends on task2 or vice versa
+                if task2 in task1.depends_on or task1 in task2.depends_on:
+                    return False
+                # Check if they share subtasks or parent relationships
+                if task1.parent == task2 or task2.parent == task1:
+                    return False
+                if task1 in task2.subtasks or task2 in task1.subtasks:
+                    return False
+    return True
+
+
 async def run_tasks_async(
     tasks: list[Task[Any]],
     thread: Thread | str | None = None,
     raise_on_failure: bool = True,
     handlers: list[Handler | AsyncHandler] | None = None,
 ) -> list[Task[Any]] | AsyncGenerator[Event, None]:
-    orchestrator = Orchestrator(
-        tasks=tasks,
-        thread=thread,
-        handlers=handlers,
-    )
-    await orchestrator.run(raise_on_failure=raise_on_failure)
-    return tasks
+    """Run tasks either concurrently (if independent) or sequentially via orchestrator."""
+    # If we have multiple independent tasks, run them concurrently
+    if len(tasks) > 1 and _tasks_are_independent(tasks):
+        # Run independent tasks concurrently using asyncio.gather
+        await asyncio.gather(*[task.run_async() for task in tasks])
+        return tasks
+    else:
+        # Use orchestrator for dependent tasks or single tasks
+        orchestrator = Orchestrator(
+            tasks=tasks,
+            thread=thread,
+            handlers=handlers,
+        )
+        await orchestrator.run(raise_on_failure=raise_on_failure)
+        return tasks
 
 
 async def run_tasks_stream(
@@ -79,6 +103,7 @@ def run_tasks(
     raise_on_failure: bool = True,
     handlers: list[Handler | AsyncHandler] | None = None,
 ) -> list[Task[Any]]:
+    """Run tasks either concurrently (if independent) or sequentially."""
     return marvin.utilities.asyncio.run_sync(
         run_tasks_async(
             tasks=tasks,